requirements.txt
DateTime==4.3
Django==2.1.10
django-pyodbc-azure==2.1.0.0
pyodbc==4.0.26
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
zope.interface==4.6.0
